requirements.txt
Flask==1.0.2
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
six==1.10.0
sh==1.12.13
python-consul==1.1.0
